+++ src/main/java/step3/domain/Car.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.domain;
 
 public class Car {
