+++ src/main/java/step3/service/GameMatcher.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.service;
 
 public interface GameMatcher {
